+++ spec/color_spec.rb
@@ -1,0 +1,8 @@
+require "helper"
+RSpec.describe Pry::Color do
+  it "raises an UnknownBrushError" do
+    expect {
+      described_class.paint "foo", :no_color
+    }.to raise_error(Pry::Color::UnknownBrushError)
+  end
+end